+++ Gopkg.toml
@@ -55,3 +55,7 @@
 [[constraint]]
   name = "github.com/go-kit/kit"
   version = "0.6.0"
+
+[[constraint]]
+  name = "github.com/sirupsen/logrus"
+  version = "1.0.5"
