+++ cli/server/server.go
@@ -3,7 +3,9 @@ package server
 import (
 	"strings"
 
+	"github.com/CityOfZion/neo-go/pkg/core"
 	"github.com/CityOfZion/neo-go/pkg/network"
+	"github.com/CityOfZion/neo-go/pkg/util"
 	"github.com/urfave/cli"
 )
 
@@ -18,6 +20,7 @@ func NewCommand() cli.Command {
 			cli.IntFlag{Name: "rpc"},
 			cli.BoolFlag{Name: "relay, r"},
 			cli.StringFlag{Name: "seed"},
+			cli.StringFlag{Name: "chain"},
 			cli.BoolFlag{Name: "privnet, p"},
 			cli.BoolFlag{Name: "mainnet, m"},
 			cli.BoolFlag{Name: "testnet, t"},
@@ -42,9 +45,38 @@ func startServer(ctx *cli.Context) error {
 		Relay:     ctx.Bool("relay"),
 	}
 
-	s := network.NewServer(cfg)
+	chain, err := newBlockchain(net, ctx.String("chain"))
+	if err != nil {
+		panic(err)
+	}
+
+	s := network.NewServer(cfg, chain)
 	s.Start()
 	return nil
+}
+
+func newBlockchain(net network.NetMode, path string) (*core.Blockchain, error) {
+	var startHash util.Uint256
+	if net == network.ModePrivNet {
+		startHash = core.GenesisHashPrivNet()
+	}
+	if net == network.ModeTestNet {
+		startHash = core.GenesisHashTestNet()
+	}
+	if net == network.ModeMainNet {
+		startHash = core.GenesisHashMainNet()
+	}
+
+	// Hardcoded for now.
+	store, err := core.NewLevelDBStore(path, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return core.NewBlockchain(
+		store,
+		startHash,
+	), nil
 }
 
 func parseSeeds(s string) []string {
