--- pkg/network/node_test.go
@@ -1,7 +1,0 @@
-package network
-
-import "testing"
-
-func TestHandleVersion(t *testing.T) {
-
-}
