--- pkg/network/p2p/message_test.go
@@ -1,1 +1,0 @@
-package p2p
