--- pkg/network/p2p/payload/payload.go
@@ -1,9 +1,0 @@
-package payload
-
-import "io"
-
-// Payload is anything that can be binary encoded/decoded.
-type Payload interface {
-    EncodeBinary(io.Writer) error
-    DecodeBinary(io.Reader) error
-}
