+++ pkg/network/peer.go
@@ -5,13 +5,10 @@ import (
 	"github.com/CityOfZion/neo-go/pkg/util"
 )
 
-// A Peer is the local representation of a remote peer.
-// It's an interface that may be backed by any concrete
-// transport.
 type Peer interface {
-	Version() *payload.Version
 	Endpoint() util.Endpoint
-	Send(*Message)
+	Disconnect(error)
+	Send(msg *Message)
 	Done() chan struct{}
-	Disconnect(err error)
+	Version() *payload.Version
 }
