+++ pkg/network/tcp_peer.go
@@ -1,70 +1,50 @@
 package network
 
 import (
-	"bytes"
 	"net"
-	"os"
 	"sync"
-	"time"
 
 	"github.com/CityOfZion/neo-go/pkg/network/payload"
 	"github.com/CityOfZion/neo-go/pkg/util"
-	log "github.com/go-kit/kit/log"
 )
 
 // TCPPeer represents a connected remote node in the
 // network over TCP.
 type TCPPeer struct {
-	// The endpoint of the peer.
+	// underlying TCP connection.
+	conn     net.Conn
 	endpoint util.Endpoint
 
-	// underlying connection.
-	conn net.Conn
-
-	// The version the peer declared when connecting.
+	// The version of the peer.
 	version *payload.Version
 
-	// connectedAt is the timestamp the peer connected to
-	// the network.
-	connectedAt time.Time
+	done   chan struct{}
+	closed chan struct{}
+	disc   chan error
 
-	// handleProto is the handler that will handle the
-	// incoming message along with its peer.
-	handleProto protoHandleFunc
-
-	// Done is used to broadcast that this peer has stopped running
-	// and should be removed as reference.
-	done chan struct{}
-
-	// Every send to this channel will terminate the Peer.
-	discErr chan error
-	closed  chan struct{}
-	wg      sync.WaitGroup
-
-	logger log.Logger
+	wg sync.WaitGroup
 }
 
-// NewTCPPeer creates a new peer from a TCP connection.
-func NewTCPPeer(conn net.Conn, fun protoHandleFunc) *TCPPeer {
-	e := util.NewEndpoint(conn.RemoteAddr().String())
-	logger := log.NewLogfmtLogger(os.Stderr)
-	logger = log.With(logger, "component", "peer", "endpoint", e)
-
+func NewTCPPeer(conn net.Conn, proto chan protoTuple) *TCPPeer {
 	return &TCPPeer{
-		endpoint:    e,
-		conn:        conn,
-		done:        make(chan struct{}),
-		logger:      logger,
-		connectedAt: time.Now().UTC(),
-		handleProto: fun,
-		discErr:     make(chan error),
-		closed:      make(chan struct{}),
+		conn:     conn,
+		done:     make(chan struct{}),
+		closed:   make(chan struct{}),
+		disc:     make(chan error),
+		endpoint: util.NewEndpoint(conn.RemoteAddr().String()),
 	}
 }
 
-// Version implements the Peer interface.
-func (p *TCPPeer) Version() *payload.Version {
-	return p.version
+// Send implements the Peer interface. This will encode the message
+// to the underlying connection. The peer will disconnect if encode
+// failed.
+func (p *TCPPeer) Send(msg *Message) {
+	if err := msg.encode(p.conn); err != nil {
+		select {
+		case p.disc <- err:
+		case <-p.closed:
+		}
+	}
 }
 
 // Endpoint implements the Peer interface.
@@ -72,57 +52,19 @@ func (p *TCPPeer) Endpoint() util.Endpoint {
 	return p.endpoint
 }
 
-// Send implements the Peer interface.
-func (p *TCPPeer) Send(msg *Message) {
-	buf := new(bytes.Buffer)
-	if err := msg.encode(buf); err != nil {
-		p.discErr <- err
-		return
-	}
-	if _, err := p.conn.Write(buf.Bytes()); err != nil {
-		p.discErr <- err
-		return
-	}
-}
-
-// Done implemnets the Peer interface. It use is to
-// notify the Node that this peer is no longer available
-// for sending messages to.
+// Done implements the Peer interface and notifies
+// all other resources operating on it to release this
+// peer.
 func (p *TCPPeer) Done() chan struct{} {
 	return p.done
 }
 
-// Disconnect terminates the peer connection.
-func (p *TCPPeer) Disconnect(err error) {
-	select {
-	case p.discErr <- err:
-	case <-p.closed:
-	}
+// Version implements the Peer interface.
+func (p *TCPPeer) Version() *payload.Version {
+	return p.version
 }
 
-func (p *TCPPeer) run() (err error) {
-	p.wg.Add(1)
-	go p.readLoop()
-
-run:
-	for {
-		select {
-		case err = <-p.discErr:
-			break run
-		}
-	}
-
-	p.conn.Close()
-	close(p.closed)
-	// Close done instead of sending empty struct.
-	// It could happen that startProtocol in Node never happens
-	// on connection errors for example.
-	close(p.done)
-	p.wg.Wait()
-	return err
-}
-
-func (p *TCPPeer) readLoop() {
+func (p *TCPPeer) readLoop(proto chan protoTuple, readErr chan error) {
 	defer p.wg.Done()
 	for {
 		select {
@@ -131,23 +73,54 @@ func (p *TCPPeer) readLoop() {
 		default:
 			msg := &Message{}
 			if err := msg.decode(p.conn); err != nil {
-				p.discErr <- err
+				readErr <- err
 				return
 			}
-			p.handleMessage(msg)
+			p.handleMessage(msg, proto)
 		}
 	}
 }
 
-func (p *TCPPeer) handleMessage(msg *Message) {
-	switch msg.CommandType() {
-	case CMDVersion:
-		version := msg.Payload.(*payload.Version)
-		p.version = version
-		fallthrough
-	default:
-		if err := p.handleProto(msg, p); err != nil {
-			p.discErr <- err
+func (p *TCPPeer) handleMessage(msg *Message, proto chan protoTuple) {
+	switch payload := msg.Payload.(type) {
+	case *payload.Version:
+		p.version = payload
+	}
+	proto <- protoTuple{
+		msg:  msg,
+		peer: p,
+	}
+}
+
+func (p *TCPPeer) run(proto chan protoTuple) (err error) {
+	readErr := make(chan error, 1)
+	p.wg.Add(1)
+	go p.readLoop(proto, readErr)
+
+run:
+	for {
+		select {
+		case err = <-p.disc:
+			break run
+		case err = <-readErr:
+			break run
 		}
 	}
+
+	// If the peer has not started the protocol with the server
+	// there will be noone reading from this channel.
+	select {
+	case p.done <- struct{}{}:
+	default:
+	}
+
+	close(p.closed)
+	p.conn.Close()
+	p.wg.Wait()
+	return
+}
+
+// Disconnect implements the Peer interface.
+func (p *TCPPeer) Disconnect(reason error) {
+	p.disc <- reason
 }
